Refactor parseRpcBalance into helpers
